By Rama krishna

+++ Master/AirlineMangementsystem/src/main/java/com/dxc/airline/repository/AirLineRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.dxc.airline.model.AirLine;
 
-public interface AirLineRepository extends JpaRepository<AirLine, Integer> {
+public interface AirLineRepository extends JpaRepository<AirLine, Long> {
 
 }
